Remove the need for the NMSUtil class.
Checking the version number of the server is simple, and doesn't require NMS. This change allows you to remove the NMSUtil class. Furthermore, instead of getting the version number and processing the package string every single time you use a sound, you should only process it once and get the version int when the class is loaded. That way you only need to compare the final int value.